Pack Textures into atlas, upload atlas to gpu, move all wgpu references to client_app, delete TextureObject and add Sprite
All textures are specified from the atlas file for an atlas, packed into a rectangle of sides of powers of two, then copied into an image.
A missing texture is automatically generated and applied to the game atlas.
Game atlas is uploaded to the gpu as the terrain and item atlas.
Wgpu dependency deleted from the library.
All wgpu references and gpu dependent code are moved to client_app subcrate, and all references updated accordingly.
TextureObject struct removed in favor of a new struct: Sprite, which defines a view in a parent image